// cypress/support/enums/add-candidate-data-enums.ts
export enum CREATE_CANDIDATE{
    FIRST_NAME = "Amal",
    MIDDLE_NAME = "Isam",
    LAST_NAME = "butmah",
    EMAIL = "amal.butmah@example.com",
    CONTACT_NUMBER = "1234567890",
    KEYWORDS = "Software Engineer",
    DATE_OF_APPLICATION = "2025-10-05",
    NOTES = "Experienced in Cypress testing",
}