// cypress/support/enums/candidate-enums.ts
export enum CANDIDATE_STATUS {
    SHORTLISTED = "Shortlisted",
    INTERVIEW_SCHEDULED = "Interview Scheduled",
    INTERVIEW_PASSED = "Interview Passed",
    REJECTED = "Rejected"
}

export enum CANDIDATE_ACTION_BUTTONS {
    SHORTLIST = "Shortlist",
    SCHEDULE_INTERVIEW = "Schedule Interview",
    MARK_INTERVIEW_PASSED = "Mark Interview Passed",
}